+++ Options/EventDelayTimerOptions.lua
@@ -100,6 +100,18 @@ local options = {
             step = 1,
             width = sliderWidth,
         },
+        PLAYER_CASTING_EMPOWERED_UPDATE = {
+            order = 9,
+            name = L["player_casting_empowered_update_delay_timer_name"],
+            desc = L["event_delay_timer_desc"],
+            type = "range",
+            get = "GetEventDelayTimerStaus",
+            set = "SetEventDelayTimerStaus",
+            min = 0,
+            softMax = 10,
+            step = 1,
+            width = sliderWidth,
+        },
     },
 }